requirements.txt
protobuf==3.19.3
streamlit==1.3.1
toml==0.10.2
google.cloud < 0.34.0
google.cloud.bigquery < 2.24.0
google.cloud.core < 1.7.2
google.auth < 1.34.0
google.resumable.media < 1.3.3
google.api.core < 1.31.1
sentence_transformers < 2.0.0
transformers < 4.6.1
# pytorch < 1.8.1
 